\name{Biostrings internals}

% XStringCodec class and methods:
\alias{class:XStringCodec}
\alias{XStringCodec-class}
\alias{XStringCodec}

\alias{initialize,XStringCodec-method}

% seqtype.R (put aliases for specific methods in corresponding man pages):
\alias{seqtype}
\alias{seqtype<-}
\alias{xscodes}
\alias{xscodes,ANY-method}
\alias{get_seqtype_conversion_lookup}

% XString-class.R:
\alias{extract_character_from_XString_by_positions}
\alias{extract_character_from_XString_by_positions,XString-method}
\alias{extract_character_from_XString_by_ranges}
\alias{extract_character_from_XString_by_ranges,XString-method}
\alias{make_XString_from_string}
\alias{make_XString_from_string,XString-method}

% XStringSet-class.R:
\alias{make_XStringSet_from_strings}
\alias{make_XStringSet_from_strings,XStringSet-method}


\title{Biostrings internals}

\description{
  Biostrings objects, classes and methods that are not intended to
  be used directly.
}

\author{H. Pagès}

\keyword{internal}
