\name{BOC_SubjectString-class}
\docType{class}

% Classes:
\alias{class:BOC_SubjectString}
\alias{BOC_SubjectString-class}
\alias{BOC_SubjectString}
\alias{class:BOC2_SubjectString}
\alias{BOC2_SubjectString-class}
\alias{BOC2_SubjectString}

% Initialization:
\alias{initialize,BOC_SubjectString-method}
\alias{initialize,BOC2_SubjectString-method}

% Other methods:
\alias{matchPattern,BOC_SubjectString-method}
\alias{matchPattern,BOC2_SubjectString-method}
\alias{countPattern,BOC2_SubjectString-method}


\title{BOC\_SubjectString and BOC2\_SubjectString objects}

\description{
  WARNING: BOC\_SubjectString and BOC2\_SubjectString objects are deprecated!
}

\author{H. Pagès}

\keyword{methods}
\keyword{classes}
