\name{BStringAlign-class}
\docType{class}

% Classes
\alias{class:BStringAlign}
\alias{BStringAlign-class}
\alias{BStringAlign}

% Constructor-like functions and generics:
\alias{initialize,BStringAlign-method}

% Accesor methods:
\alias{align1}
\alias{align1,BStringAlign-method}
\alias{align2}
\alias{align2,BStringAlign-method}
\alias{score}
\alias{score,BStringAlign-method}
\alias{length,BStringAlign-method}
\alias{nchar,BStringAlign-method}
\alias{alphabet,BStringAlign-method}

% Standard generic methods:
\alias{show,BStringAlign-method}


\title{The BStringAlign class}

\description{
  The \code{BStringAlign class} is a container for storing
  an alignment between 2 \code{\link{BString}} (or derived) objects.
}

\details{
  [TODO]
}

\section{Accesor methods}{
  In the code snippets below,
  \code{x} is a \code{BStringAlign} object.

  \describe{
    \item{}{
      \code{align1(x)} and \code{align2(x)}:
      [TODO: Document me]
    }
    \item{}{
      \code{score(x)}:
      [TODO: Document me]
    }
    \item{}{
      \code{length(x)} or \code{nchar(x)}:
      [TODO: Document me]
    }
    \item{}{
      \code{alphabet(x)}:
      [TODO: Document me]
    }
  }
}

\author{H. Pages}

\seealso{
  \code{\link{needwunsQS}},
  \link{BString-class},
  \link{DNAString-class},
  \link{RNAString-class},
  \link{AAString-class}
}

\examples{
}

\keyword{methods}
\keyword{classes}
