\name{BStringGenome}
\docType{class}

% Classes
\alias{class:BStringGenome}
\alias{BStringGenome-class}

% Constructor-like functions and generics:
\alias{initialize,BStringGenome-method}

% Standard generic methods:
\alias{show,BStringGenome-method}
\alias{$,BStringGenome-method}

% Other functions and generics:
\alias{unload}
\alias{unload,BStringGenome-method}


\title{The BStringGenome class}

\description{
  A container for the complete genome sequence of a given specie.
}

\details{
    [TODO: Put some details here]
}

\section{Standard generic methods}{
  In the code snippets below,
  \code{x} is a BStringGenome object
  and \code{name} is the name of a chromosome (character-string).

  \describe{
    \item{}{
      \code{x$name}:
      [TODO: Document me]
    }
  }
}

\section{Other functions and generics}{
  In the code snippets below,
  \code{x} is a BStringGenome object
  and \code{name} is the name of a chromosome (character-string).

  \describe{
    \item{}{
      \code{unload(x, name)}:
      [TODO: Document me]
    }
  }
}


\author{H. Pages}

\seealso{
  \link{BStringViews}, \link{DNAString}
}

\examples{
  ## [TODO: Put some examples here]
}

\keyword{methods}
\keyword{classes}
