\name{Biostrings internals}
\docType{class}

\alias{debug_bbuf}
\alias{debug_shiftor}
\alias{debug_utils}

\alias{show,externalptr-method}

\alias{class:bbuf}
\alias{bbuf-class}
\alias{bbuf}

\alias{length,bbuf-method}
\alias{toString,bbuf-method}
\alias{initialize,bbuf-method}
\alias{show,bbuf-method}
\alias{==,bbuf,bbuf-method}
\alias{!=,bbuf,bbuf-method}
\alias{[,bbuf-method}
\alias{[<-,bbuf-method}

\alias{class:cbuf}
\alias{cbuf-class}
\alias{cbuf}

\alias{show,cbuf-method}
\alias{[,cbuf-method}

\alias{class:ibuf}
\alias{ibuf-class}
\alias{ibuf}

\alias{length,ibuf-method}
\alias{initialize,ibuf-method}
\alias{show,ibuf-method}
\alias{==,ibuf,ibuf-method}
\alias{!=,ibuf,ibuf-method}
\alias{[,ibuf-method}
\alias{[<-,ibuf-method}

\alias{class:BStringCodec}
\alias{BStringCodec-class}
\alias{BStringCodec}

\alias{initialize,BStringCodec-method}

\title{Biostrings internals}

\description{
  Biostrings objects, classes and methods that are not intended to
  be used directly.
}

\author{H. Pages}

\keyword{methods}
\keyword{classes}
