\name{GENETIC_CODE}
\alias{GENETIC_CODE}


\title{The Standard Genetic Code}

\description{
  [TODO]
}

\details{
  [TODO]
}

\author{H. Pages}

\seealso{
  \code{\link{AMINO_ACID_CODE}},
  \code{\link{trinucleotideFrequency}},
  \code{\link{DNAString}}, \code{\link{RNAString}}, \code{\link{AAString}}
}

\examples{
  GENETIC_CODE
}

\keyword{utilities}
\keyword{data}
