\name{HNF4alpha}
\alias{HNF4alpha}

\title{Known HNF4alpha binding sequences}

\description{
  Seventy one known HNF4alpha binding sequences
}

\details{
  A \code{DNAStringSet} containing 71 known binding sequences for HNF4alpha.
}

\references{
  Ellrott, K., Yang, C., Sladek, F.M., Jiang, T. (2002)
  "Identifying transcription factor binding sites through Markov chain optimations",
  Bioinformatics, 18 (Suppl. 2), S100-S109.
}

\author{P. Aboyoun}

\examples{
  data(HNF4alpha)
  HNF4alpha
}

\keyword{datasets}
