\name{misc}

\alias{misc}

\alias{N50}


\title{Some miscellaneous stuff}

\description{
  Some miscellaneous stuff.
}

\usage{
  N50(csizes)
}

\arguments{
  \item{csizes}{
    A vector containing the contig sizes.
  }
}

\details{
}

\value{
  \code{N50}: The "N50" value (single integer).
}

\author{Nicolas Delhomme}

\seealso{
  \link{XStringSet-class}
}

\examples{
}

\keyword{methods}
