\name{moved_to_pwalign}

\alias{moved_to_pwalign}

\alias{writePairwiseAlignments}
\alias{nucleotideSubstitutionMatrix}
\alias{errorSubstitutionMatrices}
\alias{qualitySubstitutionMatrices}
\alias{insertion}
\alias{deletion}
\alias{unaligned}
\alias{aligned}
\alias{indel}
\alias{nindel}
\alias{PairwiseAlignments}
\alias{alignedPattern}
\alias{alignedSubject}
\alias{PairwiseAlignmentsSingleSubject}
\alias{nedit}
\alias{mismatchTable}
\alias{mismatchSummary}
\alias{compareStrings}
\alias{pid}
\alias{pairwiseAlignment}
\alias{stringDist}


\title{InDel objects}

\description{
  Starting with BioC 3.19, the following functions are defined in the
  \pkg{pwalign} package:
  \itemize{
      \item \code{writePairwiseAlignments}
      \item \code{nucleotideSubstitutionMatrix}
      \item \code{errorSubstitutionMatrices}
      \item \code{qualitySubstitutionMatrices}
      \item \code{insertion}
      \item \code{deletion}
      \item \code{unaligned}
      \item \code{aligned}
      \item \code{indel}
      \item \code{nindel}
      \item \code{PairwiseAlignments}
      \item \code{alignedPattern}
      \item \code{alignedSubject}
      \item \code{PairwiseAlignmentsSingleSubject}
      \item \code{nedit}
      \item \code{mismatchTable}
      \item \code{mismatchSummary}
      \item \code{compareStrings}
      \item \code{pid}
      \item \code{pairwiseAlignment}
      \item \code{stringDist}
  }
}

