\name{predefined_scoring_matrices}
\docType{data}

\alias{predefined_scoring_matrices}
\alias{BLOSUM45}
\alias{BLOSUM50}
\alias{BLOSUM62}
\alias{BLOSUM80}
\alias{BLOSUM100}
\alias{PAM30}
\alias{PAM40}
\alias{PAM70}
\alias{PAM120}
\alias{PAM250}

\alias{print.moved_to_pwalign_pkg}

\title{Predefined scoring matrices}

\description{
  Predefined scoring matrices for nucleotide and amino acid alignments.

  WARNING: All the BLOSUM* and PAM* scoring matrices listed below are
  now located in the \pkg{pwalign} package and will soon be removed from
  the \pkg{Biostrings} package.
}

\usage{
data(BLOSUM45)
data(BLOSUM50)
data(BLOSUM62)
data(BLOSUM80)
data(BLOSUM100)
data(PAM30)
data(PAM40)
data(PAM70)
data(PAM120)
data(PAM250)
}

\format{
  See \code{?pwalign::\link[pwalign]{predefined_scoring_matrices}}
  in the \pkg{pwalign} package.
}

\details{
  See \code{?pwalign::\link[pwalign]{predefined_scoring_matrices}}
  in the \pkg{pwalign} package.
}

\examples{
## See ?pwalign::predefined_scoring_matrices in the pwalign package.
}

\keyword{data}
\keyword{datasets}
