\name{reverseComplement}

\alias{reverseComplement}
\alias{reverseComplement,DNAString-method}
\alias{reverseComplement,RNAString-method}

\alias{reverseComplement}


\title{Function to reverse a DNA or RNA sequence and complement each base}

\description{
  [TODO]
}

\usage{
  reverseComplement(x)
}

\arguments{
  \item{x}{
    A \link{DNAString} or \link{RNAString} object.
  }
}

\details{
}

\value{
}

\author{H. Pages}

\seealso{}

\examples{
}

\keyword{classes}

