\name{WCP}
\docType{class}

\alias{class:WCP}
\alias{WCP-class}
\alias{WCP}
\alias{class:B_WCP}
\alias{B_WCP-class}
\alias{B_WCP}
\alias{class:DNA_WCP}
\alias{DNA_WCP-class}
\alias{DNA_WCP}
\alias{class:RNA_WCP}
\alias{RNA_WCP-class}
\alias{RNA_WCP}
\alias{class:AA_WCP}
\alias{AA_WCP-class}
\alias{AA_WCP}

\alias{maxWeights,WCP-method}
\alias{show,WCP-method}


\title{Weighted Clustered Positions (WCP) objects}

\description{
  The WCP class is now defunct.
}

\author{P. Aboyoun}

\seealso{
  \code{\link{matchWCP}},
  \link{XString-class}
}

\keyword{methods}
\keyword{classes}
