\name{XKeySortedData}
\docType{class}

\alias{class:XKeySortedData}
\alias{XKeySortedData-class}
\alias{XKeySortedData}
\alias{class:BKeySortedData}
\alias{BKeySortedData-class}
\alias{BKeySortedData}
\alias{class:DNAKeySortedData}
\alias{DNAKeySortedData-class}
\alias{DNAKeySortedData}
\alias{class:RNAKeySortedData}
\alias{RNAKeySortedData-class}
\alias{RNAKeySortedData}
\alias{class:AAKeySortedData}
\alias{AAKeySortedData-class}
\alias{AAKeySortedData}

\alias{[,XKeySortedData-method}
\alias{[[,XKeySortedData-method}
\alias{dataKey}
\alias{dataKey,XKeySortedData-method}
\alias{dataTable}
\alias{dataTable,XKeySortedData-method}
\alias{dim,XKeySortedData-method}
\alias{dimnames,XKeySortedData-method}
\alias{length,XKeySortedData-method}
\alias{xsbasetype,XKeySortedData-method}


\title{Data Dictionaries with XString-based Keys}

\description{
  The XKeySortedData class is a container for storing a dictionary with
  XString-based keys and DataFrame (an IRanges class) values.
}

\author{P. Aboyoun}

\seealso{
  \link{XStringSet-class},
  \link[IRanges]{DataFrame-class}
}

\keyword{methods}
\keyword{classes}
