\name{XKeySortedDataList}
\docType{class}

\alias{class:XKeySortedDataList}
\alias{XKeySortedDataList-class}
\alias{XKeySortedDataList}
\alias{class:BKeySortedDataList}
\alias{BKeySortedDataList-class}
\alias{BKeySortedDataList}
\alias{class:DNAKeySortedDataList}
\alias{DNAKeySortedDataList-class}
\alias{DNAKeySortedDataList}
\alias{class:RNAKeySortedDataList}
\alias{RNAKeySortedDataList-class}
\alias{RNAKeySortedDataList}
\alias{class:AAKeySortedDataList}
\alias{AAKeySortedDataList-class}
\alias{AAKeySortedDataList}

\alias{seqtype,XKeySortedDataList-method}


\title{List of Data Dictionaries with XString-based Keys}

\description{
  WARNING: The XKeySortedDataList class and subclasses are defunct.

  The XKeySortedDataList class is a container for storing a list of
  dictionaries with XString-based keys and DataFrame (an IRanges class) values.
}

\author{P. Aboyoun}

\seealso{
  \link{XKeySortedData-class},
  \link[IRanges]{SimpleList-class}
}

\keyword{methods}
\keyword{classes}
