\name{yeastSEQCHR1}
\alias{yeastSEQCHR1}
\title{An annotation data file for CHR1 in the yeastSEQ package}
\description{
This is a single character string containing DNA sequence of yeast chromosome number 1.  The data were obtained from the Saccharomyces Genome Database
(\url{ftp://genome-ftp.stanford.edu/pub/yeast/data\_download/sequence/genomic\_sequence/chromosomes/fasta/}).
}
\details{
Annotation based on data provided by Yeast Genome project.

Source data built:Yeast Genome data are built at various time intervals. Sources used were downloaded Fri Nov 21 14:00:47 2003
Package built: Fri Nov 21 14:00:47 2003
}
\references{
\url{http://www.yeastgenome.org/DownloadContents.shtml}
}
\examples{
data(yeastSEQCHR1)
nchar(yeastSEQCHR1)
}
\keyword{datasets}
