\name{alphabetFrequency}

\alias{alphabetFrequency}
\alias{alphabetFrequency,DNAString-method}
\alias{alphabetFrequency,RNAString-method}
\alias{alphabetFrequency,BStringViews-method}

\alias{alphabetFrequency}


\title{Function to calculate the frequency of letters in a biological sequence}

\description{
  [TODO]
}

\usage{
  alphabetFrequency(x, baseOnly=TRUE)
}

\arguments{
  \item{x}{
    A \link{DNAString},  \link{RNAString} or \link{BStringViews} object.
  }
  \item{baseOnly}{
    \code{TRUE} or \code{FALSE}.
  }
}

\details{
}

\value{
}

\author{H. Pages}

\seealso{}

\examples{
}

\keyword{classes}

