\name{as.complex}

% Standard generic methods:
\alias{as.complex,DNAString-method}

\title{Converting a DNAString to a vector of complexes}

\description{
}

\usage{
  as.complex(x, base_values)
}

\details{
}

\author{H. Pages}

\seealso{
  \link{DNAString}
}

\examples{
}

\keyword{methods}
\keyword{classes}
