\name{as.complex}

% Standard generic methods:
\alias{as.complex}
\alias{as.complex,DNAString-method}

\title{Converting a DNAString to a vector of complexes}

\description{
}

\usage{
  as.complex(x, ...)
}

\arguments{
  \item{x}{
    A \code{\link{DNAString}} object.
  }
  \item{...}{
    A named complex vector \code{baseValues} containing the values
    associated to each base.
  }
}

\details{
}

\author{H. Pages}

\seealso{
  \link{DNAString}
}

\examples{
}

\keyword{methods}

