\name{Biostrings internals}
\docType{class}

% SparseList class and methods:
\alias{class:SparseList}
\alias{SparseList-class}
\alias{SparseList}

\alias{length,SparseList-method}
%\alias{ls,SparseList-method}
\alias{as.list,SparseList-method}
\alias{[[,SparseList-method}

% XStringCodec class and methods:
\alias{class:XStringCodec}
\alias{XStringCodec-class}
\alias{XStringCodec}

\alias{initialize,XStringCodec-method}

% seqtype.R (put aliases for specific methods in corresponding man pages):
\alias{seqtype}
\alias{seqtype<-}
\alias{get_seqtype_conversion_lookup}

% XString-class.R:
\alias{make_XString_from_string}
\alias{make_XString_from_string,XString-method}

% XStringSet-class.R:
\alias{make_XStringSet_from_strings}
\alias{make_XStringSet_from_strings,XStringSet-method}

\title{Biostrings internals}

\description{
  Biostrings objects, classes and methods that are not intended to
  be used directly.
}

\author{H. Pagès}

\keyword{internal}
\keyword{methods}
\keyword{classes}
