\name{Biostrings internals}
\docType{class}

% Dups class and methods:
\alias{class:Dups}
\alias{Dups-class}
\alias{Dups}

\alias{length,Dups-method}
\alias{duplicated,Dups-method}
\alias{show,Dups-method}

% SparseList class and methods:
\alias{class:SparseList}
\alias{SparseList-class}
\alias{SparseList}

\alias{length,SparseList-method}
\alias{ls,SparseList-method}
\alias{as.list,SparseList-method}
\alias{[[,SparseList-method}

% XRaw class and methods:
\alias{show,externalptr-method}

\alias{class:XRaw}
\alias{XRaw-class}
\alias{XRaw}

\alias{length,XRaw-method}
\alias{as.integer,XRaw-method}
\alias{toString,XRaw-method}
\alias{initialize,XRaw-method}
\alias{show,XRaw-method}
\alias{==,XRaw,XRaw-method}
\alias{!=,XRaw,XRaw-method}
\alias{[,XRaw-method}
\alias{[<-,XRaw-method}

\alias{class:PrintableXRaw}
\alias{PrintableXRaw-class}
\alias{PrintableXRaw}

\alias{show,PrintableXRaw-method}
\alias{[,PrintableXRaw-method}

%XStringCodec class and methods:
\alias{class:XStringCodec}
\alias{XStringCodec-class}
\alias{XStringCodec}

\alias{initialize,XStringCodec-method}
\alias{alphabet,XStringCodec-method}


\title{Biostrings internals}

\description{
  Biostrings objects, classes and methods that are not intended to
  be used directly.
}

\author{H. Pages}

\keyword{internal}
\keyword{methods}
\keyword{classes}
