\name{Biostrings internals}
\docType{class}

% Dups class and methods:
\alias{class:Dups}
\alias{Dups-class}
\alias{Dups}

\alias{length,Dups-method}
\alias{duplicated,Dups-method}
\alias{show,Dups-method}

% SparseList class and methods:
\alias{class:SparseList}
\alias{SparseList-class}
\alias{SparseList}

\alias{length,SparseList-method}
\alias{ls,SparseList-method}
\alias{as.list,SparseList-method}
\alias{[[,SparseList-method}

%XStringCodec class and methods:
\alias{class:XStringCodec}
\alias{XStringCodec-class}
\alias{XStringCodec}

\alias{initialize,XStringCodec-method}
\alias{alphabet,XStringCodec-method}


\title{Biostrings internals}

\description{
  Biostrings objects, classes and methods that are not intended to
  be used directly.
}

\author{H. Pages}

\keyword{internal}
\keyword{methods}
\keyword{classes}
