\name{Biostrings internals}
\docType{class}

\alias{class:SparseList}
\alias{SparseList-class}
\alias{SparseList}

\alias{length,SparseList-method}
\alias{ls,SparseList-method}
\alias{as.list,SparseList-method}
\alias{[[,SparseList-method}

\alias{show,externalptr-method}

\alias{class:XRaw}
\alias{XRaw-class}
\alias{XRaw}

\alias{length,XRaw-method}
\alias{as.integer,XRaw-method}
\alias{toString,XRaw-method}
\alias{initialize,XRaw-method}
\alias{show,XRaw-method}
\alias{==,XRaw,XRaw-method}
\alias{!=,XRaw,XRaw-method}
\alias{[,XRaw-method}
\alias{[<-,XRaw-method}

\alias{class:PrintableXRaw}
\alias{PrintableXRaw-class}
\alias{PrintableXRaw}

\alias{show,PrintableXRaw-method}
\alias{[,PrintableXRaw-method}

\alias{class:XInteger}
\alias{XInteger-class}
\alias{XInteger}

\alias{length,XInteger-method}
\alias{as.integer,XInteger-method}
\alias{initialize,XInteger-method}
\alias{show,XInteger-method}
\alias{==,XInteger,XInteger-method}
\alias{!=,XInteger,XInteger-method}
\alias{[,XInteger-method}
\alias{[<-,XInteger-method}

\alias{class:XNumeric}
\alias{XNumeric-class}
\alias{XNumeric}

\alias{length,XNumeric-method}
\alias{as.numeric,XNumeric-method}
\alias{initialize,XNumeric-method}
\alias{show,XNumeric-method}
\alias{==,XNumeric,XNumeric-method}
\alias{!=,XNumeric,XNumeric-method}
\alias{[,XNumeric-method}
\alias{[<-,XNumeric-method}

\alias{class:XStringCodec}
\alias{XStringCodec-class}
\alias{XStringCodec}

\alias{initialize,XStringCodec-method}

\alias{alphabet,XStringCodec-method}


\title{Biostrings internals}

\description{
  Biostrings objects, classes and methods that are not intended to
  be used directly.
}

\author{H. Pages}

\keyword{internal}
\keyword{methods}
\keyword{classes}
