\name{Biostrings internals}
\docType{class}

\alias{debug_CharBuffer}
\alias{debug_shiftor}
\alias{debug_utils}

\alias{show,externalptr-method}

\alias{class:CharBuffer}
\alias{CharBuffer-class}
\alias{CharBuffer}

\alias{length,CharBuffer-method}
\alias{toString,CharBuffer-method}
\alias{initialize,CharBuffer-method}
\alias{show,CharBuffer-method}
\alias{==,CharBuffer,CharBuffer-method}
\alias{!=,CharBuffer,CharBuffer-method}
\alias{[,CharBuffer-method}
\alias{[<-,CharBuffer-method}

\alias{class:PrintableCharBuffer}
\alias{PrintableCharBuffer-class}
\alias{PrintableCharBuffer}

\alias{show,PrintableCharBuffer-method}
\alias{[,PrintableCharBuffer-method}

\alias{class:IntBuffer}
\alias{IntBuffer-class}
\alias{IntBuffer}

\alias{length,IntBuffer-method}
\alias{initialize,IntBuffer-method}
\alias{show,IntBuffer-method}
\alias{==,IntBuffer,IntBuffer-method}
\alias{!=,IntBuffer,IntBuffer-method}
\alias{[,IntBuffer-method}
\alias{[<-,IntBuffer-method}

\alias{class:BStringCodec}
\alias{BStringCodec-class}
\alias{BStringCodec}

\alias{initialize,BStringCodec-method}

\alias{alphabet}
\alias{alphabet,BStringCodec-method}


\title{Biostrings internals}

\description{
  Biostrings objects, classes and methods that are not intended to
  be used directly.
}

\author{H. Pages}

\keyword{methods}
\keyword{classes}
