\name{Biostrings internals}
\docType{class}

\alias{debug_ByteBuffer}
\alias{debug_shiftor}
\alias{debug_utils}

\alias{show,externalptr-method}

\alias{class:ByteBuffer}
\alias{ByteBuffer-class}
\alias{ByteBuffer}

\alias{length,ByteBuffer-method}
\alias{toString,ByteBuffer-method}
\alias{initialize,ByteBuffer-method}
\alias{show,ByteBuffer-method}
\alias{==,ByteBuffer,ByteBuffer-method}
\alias{!=,ByteBuffer,ByteBuffer-method}
\alias{[,ByteBuffer-method}
\alias{[<-,ByteBuffer-method}

\alias{class:cbuf}
\alias{cbuf-class}
\alias{cbuf}

\alias{show,cbuf-method}
\alias{[,cbuf-method}

\alias{class:ibuf}
\alias{ibuf-class}
\alias{ibuf}

\alias{length,ibuf-method}
\alias{initialize,ibuf-method}
\alias{show,ibuf-method}
\alias{==,ibuf,ibuf-method}
\alias{!=,ibuf,ibuf-method}
\alias{[,ibuf-method}
\alias{[<-,ibuf-method}

\alias{class:BStringCodec}
\alias{BStringCodec-class}
\alias{BStringCodec}

\alias{initialize,BStringCodec-method}

\title{Biostrings internals}

\description{
  Biostrings objects, classes and methods that are not intended to
  be used directly.
}

\author{H. Pages}

\keyword{methods}
\keyword{classes}
