\name{Biostrings internals}
\docType{class}

% Dups class and methods:
\alias{class:Dups}
\alias{Dups-class}
\alias{Dups}

\alias{length,Dups-method}
\alias{duplicated,Dups-method}
\alias{show,Dups-method}

% SparseList class and methods:
\alias{class:SparseList}
\alias{SparseList-class}
\alias{SparseList}

\alias{length,SparseList-method}
\alias{ls,SparseList-method}
\alias{as.list,SparseList-method}
\alias{[[,SparseList-method}

% RawPtr class and methods:
\alias{show,externalptr-method}

\alias{class:RawPtr}
\alias{RawPtr-class}
\alias{RawPtr}

\alias{length,RawPtr-method}
\alias{as.integer,RawPtr-method}
\alias{toString,RawPtr-method}
\alias{initialize,RawPtr-method}
\alias{show,RawPtr-method}
\alias{==,RawPtr,RawPtr-method}
\alias{!=,RawPtr,RawPtr-method}
\alias{[,RawPtr-method}
\alias{[<-,RawPtr-method}

\alias{class:PrintableRawPtr}
\alias{PrintableRawPtr-class}
\alias{PrintableRawPtr}

\alias{show,PrintableRawPtr-method}
\alias{[,PrintableRawPtr-method}

%XStringCodec class and methods:
\alias{class:XStringCodec}
\alias{XStringCodec-class}
\alias{XStringCodec}

\alias{initialize,XStringCodec-method}
\alias{alphabet,XStringCodec-method}


\title{Biostrings internals}

\description{
  Biostrings objects, classes and methods that are not intended to
  be used directly.
}

\author{H. Pages}

\keyword{internal}
\keyword{methods}
\keyword{classes}
