\name{matchProbePair}

\alias{matchProbePair}


\title{Find probe pair matches}

\description{
  [TODO]
}

\usage{
  matchProbePair(Fprobe, Rprobe, subject, logfile=NULL, verbose=FALSE)
}

\arguments{
  \item{Fprobe}{
    The forward probe.
  }
  \item{Rprobe}{
    The reverse probe.
  }
  \item{subject}{
    A \link{DNAString} object containing the subject string.
  }
  \item{logfile}{
    A file used for logging.
  }
  \item{verbose}{
    \code{TRUE} or \code{FALSE}.
  }
}

\details{
}

\value{
}

\author{H. Pages}

\seealso{}

\examples{
}

\keyword{methods}

